feat: refatoração N-Tier — CMSAPI, CMSXRepo, ICMSX e módulo de pedidos
- Separa controllers para projeto CMSAPI (nova camada de apresentação)
- Cria CMSXRepo com implementações de repositório (nova camada de dados)
- Cria ICMSX com interfaces de repositório e agente IA
- Adiciona novos modelos: Grupo, IaCache, IaConfig, IaUso, LayoutTemplate
- Adiciona Service Bus publisher/consumer para eventos de pedido
- Cria .gitignore para o projeto

+++ CMSXData/Models/Cambio.cs
@@ -11,5 +11,5 @@ public partial class Cambio
 
     public string? Moedasxml { get; set; }
 
-    public short? Tipo { get; set; }
+    public byte? Tipo { get; set; }
 }
